Add Expiation damage mods
